+++ data/education-platform-data.js
@@ -36,7 +36,10 @@
     teacher: ['view_own_learning', 'view_own_assignments', 'view_own_feedback', 'view_teacher_dashboard', 'manage_class', 'assign_content', 'review_submission', 'build_course', 'create_organization'],
     reviewer: ['view_own_learning', 'view_own_assignments', 'view_own_feedback', 'review_content'],
     content_editor: ['view_own_learning', 'view_own_assignments', 'view_own_feedback', 'review_content', 'manage_content'],
-    admin: ['view_own_learning', 'view_own_assignments', 'view_own_feedback', 'view_teacher_dashboard', 'manage_class', 'assign_content', 'review_submission', 'build_course', 'create_organization', 'manage_organization', 'manage_content', 'review_content']
+    content_creator: ['view_own_learning', 'view_own_assignments', 'view_own_feedback', 'build_course', 'manage_content'],
+    center_admin: ['view_own_learning', 'view_own_assignments', 'view_own_feedback', 'view_teacher_dashboard', 'manage_class', 'assign_content', 'review_submission', 'build_course', 'create_organization', 'manage_organization'],
+    admin: ['view_own_learning', 'view_own_assignments', 'view_own_feedback', 'view_teacher_dashboard', 'manage_class', 'assign_content', 'review_submission', 'build_course', 'create_organization', 'manage_organization', 'manage_content', 'review_content'],
+    super_admin: ['view_own_learning', 'view_own_assignments', 'view_own_feedback', 'view_teacher_dashboard', 'manage_class', 'assign_content', 'review_submission', 'build_course', 'create_organization', 'manage_organization', 'manage_content', 'review_content']
   });
 
   global.KLEARN_EDUCATION_PLATFORM_DATA = Object.freeze({